rename to technicals.py
