+++ src/routes/__init__.py
@@ -1,7 +1,7 @@
 from .finance_news import router as finance_news_router
 from .historical_prices import router as historical_prices_router
 from .hours import router as hours_router
-from .indicators import router as indicators_router
+from .technicals import router as indicators_router
 from .indices import router as indices_router
 from .movers import router as movers_router
 from .quotes import router as quotes_router
