+++ src/routes/technicals.py
